+++ ketraterm-intellij-plugin/src/main/kotlin/io/github/ketraterm/intellij/settings/KetraTermIntellijThemePalette.kt
@@ -96,6 +96,7 @@ internal object KetraTermIntellijThemePalette {
             cursorBackground = cursorBackground,
             indexedColors = indexed,
             boldAsBright = true,
+            isDark = dark,
         )
     }